Use size_t for size field and increase file buffer size
- Change size field from int to size_t to handle large allocations
- Increase file path buffer from 80 to 256 characters
- Use %zu format specifier for size_t values
- Prevents truncation issues with large allocations
- All tests pass in both DEBUG and non-DEBUG modes

+++ src/xalloc.c
@@ -166,8 +166,8 @@ static void insert_ptr(arena_t *arena, void *p)
     arena->allocated_ptrs[idx].ptr = p;
     snprintf(arena->allocated_ptrs[idx].file, sizeof(arena->allocated_ptrs[idx].file), "%s", file);
     arena->allocated_ptrs[idx].line = line;
-    arena->allocated_ptrs[idx].size = (int)size;
-    fprintf(stderr, "ALLOC: %p %s:%d %d\n", p, file, line, (int)size);
+    arena->allocated_ptrs[idx].size = size;
+    fprintf(stderr, "ALLOC: %p %s:%d %zu\n", p, file, line, size);
 #else
     arena->allocated_ptrs[idx] = p;
 #endif
@@ -266,7 +266,7 @@ void *arena_xrealloc(arena_t *arena, void *old_ptr, size_t new_size)
     if (idx >= 0)
     {
         // Print old allocation info
-        fprintf(stderr, "REALLOC: %p %s:%d %d -> ", 
+        fprintf(stderr, "REALLOC: %p %s:%d %zu -> ", 
                 arena->allocated_ptrs[idx].ptr,
                 arena->allocated_ptrs[idx].file,
                 arena->allocated_ptrs[idx].line,
@@ -284,7 +284,7 @@ void *arena_xrealloc(arena_t *arena, void *old_ptr, size_t new_size)
         remove_ptr_at(arena, idx);
 #ifdef DEBUG
     // Print new allocation info
-    fprintf(stderr, "%p %s:%d %d\n", p, file, line, (int)new_size);
+    fprintf(stderr, "%p %s:%d %zu\n", p, file, line, new_size);
     insert_ptr(arena, p, file, line, new_size);
 #else
     insert_ptr(arena, p);
@@ -336,7 +336,7 @@ void arena_xfree(arena_t *arena, void *p)
     }
 
 #ifdef DEBUG
-    fprintf(stderr, "FREE: %p %s:%d %d -> %p (freed):0 0\n",
+    fprintf(stderr, "FREE: %p %s:%d %zu -> %p (freed):0 0\n",
             arena->allocated_ptrs[idx].ptr,
             arena->allocated_ptrs[idx].file,
             arena->allocated_ptrs[idx].line,
@@ -384,7 +384,7 @@ void arena_reset_ex(arena_t *arena, bool verbose)
     {
         for (long i = 0; i < arena->allocated_count; i++)
         {
-            fprintf(stderr, "LEAK: %p %s:%d %d\n",
+            fprintf(stderr, "LEAK: %p %s:%d %zu\n",
                     arena->allocated_ptrs[i].ptr,
                     arena->allocated_ptrs[i].file,
                     arena->allocated_ptrs[i].line,
@@ -491,7 +491,7 @@ void arena_reset(bool verbose)
     {
         for (long i = 0; i < global_arena.allocated_count; i++)
         {
-            fprintf(stderr, "LEAK: %p %s:%d %d\n",
+            fprintf(stderr, "LEAK: %p %s:%d %zu\n",
                     global_arena.allocated_ptrs[i].ptr,
                     global_arena.allocated_ptrs[i].file,
                     global_arena.allocated_ptrs[i].line,
